Rewrite thread callbacks to use mutekix_thread_wrapper
There were sporatic crashes when quitting the emulator, which might be realted to ABI compatibility issues (memory layout changes causes the stack to be misaligned, which may then lead to corrupttion of some on-stack values on high optimization). Changing it just to be safe.

Also disabled PIC since it spammed unnecessary REL32s, and increased stack memory size for worker threads just in case.