Add scheduler timer as an alternative to RTC
On some	boards,	the RTC	timer don't have the accuracy we need to properly time the frames, which makes the emulator run too fast. For those boards, we can use a high-priority thread to measure the scheduler interrupt instead. This may be slightly less accurate than true millisecond RTC but it'd be the only way to get the timing to work somewhat correctly when the RTC does not have true millisecond precision.