Try to make input more robust
On weak systems like BA110 the old DIS routine still loses track of presses when under high pressure. One theory on how to fix this is to increase the event polling rate and have a release timer that ticks down until sufficient amount of time has passed. This patch implements that. Whether this is actually useful has yet to be seen.