Make it compile with cgb branch of Peanut-GB
The cgb branch is based on an older version of Peanut-GB that has a slightly different API. Some compatibility code were added to get around this. Moreover, there's now a target Peanut-GBC-bestape that can be used to build the emulator with CGB support.

This won't work on its own. You need to either use the cgb branch of **this repository** or check out the cgb branch of Peanut-GB before building Peanut-GBC-bestape.