Rename project to WoodyBoy
This is to hopefully avoid confusion with Peanut-GB (the project) and other ports of it.

"Woody" has similar pronunciation to the Taiwanese mandarain name of Besta (and it was used in some of the online promotion material of Besta dictionaries back in 2010-ish), and I think MutekiBoy sounds really lame, so here we go.